requirements.txt
amqp==5.0.6
asgiref==3.2.3
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
CacheControl==0.12.6
cachy==0.3.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cleo==0.8.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.8
distlib==0.3.2
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
environ==1.0
filelock==3.0.12
html5lib==1.1
idna==3.2
jeepney==0.7.1
keyring==21.8.0
kombu==5.1.0
lockfile==0.12.2
msgpack==1.0.2
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pipenv==2021.5.29
pkginfo==1.7.1
platformdirs==2.3.0
poetry==1.1.8
poetry-core==1.0.4
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
PyJWT==1.7.1
pylev==1.4.0
pyparsing==2.4.7
python-environ==0.4.54
pytz==2019.3
requests==2.26.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sqlparse==0.3.0
tomlkit==0.7.2
urllib3==1.26.6
vine==5.0.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
wcwidth==0.2.5
webencodings==0.5.1
